Demonstrate working via JSON in go
The downside to the Einride protobuf Update method is that it requires
us to know a little bit more about the type of the message when we load
that message from a serialized version. (At least, as far as I can tell.
It may be possible to work around.)

JSON in Go has a nice and useful generic representation of
map[string]interface{}. This applies basically the same algorithm as the
Python dict updater to such a map in Go, with some helpers to deal with
the fact that field masks use message names (e.g. state_code) and the
protojson.Marshal output uses the protobuf JSON names (e.g. stateCode).

The results should be deserializable via protojson.Unmarshal as well as
json.Unmarshal.